[F #1] Perform short character list request

+++ models/stats.go
@@ -1,10 +1,10 @@
 package models
 
 type Stats struct {
-	HP          int     `json:"hp"`
-	MP          int     `json:"mp"`
-	MaxHP       int     `json:"maxHp"`
-	MaxMP       int     `json:"maxMp"`
+	HP          float32 `json:"hp"`
+	MP          float32 `json:"mp"`
+	MaxHP       float32 `json:"maxHp"`
+	MaxMP       float32 `json:"maxMp"`
 	HitChanse   float32 `json:"hitChanse"`
 	CritChanse  float32 `json:"critChanse"`
 	DodgeChanse float32 `json:"dodgeChance"`
